% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieIndexes.R
\name{getBowtieIndexesDir}
\alias{getBowtieIndexesDir}
\title{Get bowtie indexes directory}
\usage{
getBowtieIndexesDir()
}
\value{
If the environment path `BOWTIE_INDEXES` is found,
    \code{getBowtieDir} will return the directory name.
}
\description{
Return bowtie indexes directory name if set up by user.
}
\examples{
getBowtieIndexesDir()

}
\author{
Jean-Philippe Fortin
}
