% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieIndexes.R
\name{listBowtieIndexes}
\alias{listBowtieIndexes}
\title{List available bowtie indexes}
\usage{
listBowtieIndexes()
}
\value{
Character vector ofall available bowtie indexes found
in the directory returned by \code{getBowtieIndexesDir}.
}
\description{
List all available bowtie indexes found in
    the directory returned by \code{getBowtieIndexesDir}.
}
\examples{
listBowtieIndexes()

}
\author{
Jean-Philippe Fortin
}
