% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieIndexes.R
\name{setBowtieIndexesDir}
\alias{setBowtieIndexesDir}
\title{Set bowtie indexes directory}
\usage{
setBowtieIndexesDir(dir)
}
\arguments{
\item{dir}{String specifying directory where bowtie indexes are stored.}
}
\value{
No object is returned. Instead, the BOWTIE_INDEXES
    environment variable is set to 'dir' for the remaining of the session.
}
\description{
Set bowtie indexes directory in an R session.
}
\examples{
setBowtieIndexesDir("./")

}
\author{
Jean-Philippe Fortin
}
