% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBowtie.R
\name{runBowtie}
\alias{runBowtie}
\title{Run bowtie short-read aligner (ungapped alignments)}
\usage{
runBowtie(
  sequences,
  bowtie_index = NULL,
  n_mismatches = 0,
  all_alignments = TRUE,
  n_max_alignments = 1000,
  verbose = TRUE
)
}
\arguments{
\item{sequences}{Character vector of DNA sequences.}

\item{bowtie_index}{Path to the bowtie genome index to be used
for alignment.}

\item{n_mismatches}{Integer between 0 and 3 specifying maximum number
of mismatches allowed between the query sequences and the
index sequences.}

\item{all_alignments}{Should all possible alignments be returned?
TRUE by default.}

\item{n_max_alignments}{Maximum number of alignments to return if
\code{all_alignments} is FALSE. 1000 by default.}

\item{verbose}{Should messages be printed to the consolde? TRUE by default.}
}
\value{
\strong{runBowtie} returns sequence alignment data, 
    including genomic coordinates and sequence, and position of 
    mismatches relative to the sstart of the sequence.
}
\description{
Return bowtie alignments for a list of short sequences
    for a prebuilt bowtie index.
}
\details{
\code{runBowtie} can be used to search for any short read
    sequence in a reference genome. To search for sequences while imposing
    constraints on PAM sequences (such as gRNA spacer sequences), see
    \code{runCrisprBowtie} instead.
}
\examples{
fasta <- system.file(package="crisprBowtie", "example/chr1.fa")
outdir <- tempdir()
Rbowtie::bowtie_build(fasta,outdir=outdir, force=TRUE, prefix="tempIndex")
index <- file.path(outdir, "tempIndex")
seqs <- c("GGAAGT",
          "GTGGAC",
          "GTGTGC") 
results <- runBowtie(seqs, bowtie_index=index, n_mismatches=2)

}
\author{
Jean-Philippe Fortin
}
